Use local repositories for testing.
This is a bit (well, a lot) heavier on the disk and is a bit more
complex but doesn't require anyone to hit the network every time
they want to run tests.

Bonus: I don't have to maintain accounts at a handful of different,
and otherwise useless, services.

There's a lot of ugly duplication in the Makefile which will need to
be cleaned up. Perhaps move this step into the tests themselves?